@@ -19,8 +19,10 @@ from pysatl_criterion.statistics.normal import (
     GlenLeemisBarrNormalityGofStatistic,
     GMGNormalityGofStatistic,
     GraphAverageDegreeNormalityGofStatistic,
+    GraphCliqueNumberNormalityGofStatistic,
     GraphConnectedComponentsNormalityGofStatistic,
     GraphEdgesNumberNormalityGofStatistic,
+    GraphIndependenceNumberNormalityGofStatistic,
     GraphMaxDegreeNormalityGofStatistic,
     Hosking1NormalityGofStatistic,
     Hosking2NormalityGofStatistic,
@@ -801,7 +803,10 @@ def test_graph_edges_number_normality_criterion(data, result):
 
 
 def test_graph_edges_number_normality_criterion_code():
-    assert "EdgesNumber_NORMALITY_GOODNESS_OF_FIT" == GraphEdgesNumberNormalityGofStatistic().code()
+    assert (
+        "EDGESNUMBER_GRAPH_NORMALITY_GOODNESS_OF_FIT"
+        == GraphEdgesNumberNormalityGofStatistic().code()
+    )
 
 
 @pytest.mark.parametrize(
@@ -817,7 +822,9 @@ def test_graph_max_degree_normality_criterion(data, result):
 
 
 def test_graph_max_degree_normality_criterion_code():
-    assert "MaxDegree_NORMALITY_GOODNESS_OF_FIT" == GraphMaxDegreeNormalityGofStatistic().code()
+    assert (
+        "MAXDEGREE_GRAPH_NORMALITY_GOODNESS_OF_FIT" == GraphMaxDegreeNormalityGofStatistic().code()
+    )
 
 
 @pytest.mark.parametrize(
@@ -834,8 +841,46 @@ def test_mean_vertex_degree_criterion(data, result):
 
 def test_mean_vertex_degree_criterion_code():
     assert (
-        "AverageDegree_NORMALITY_GOODNESS_OF_FIT"
+        "AVGDEGREE_GRAPH_NORMALITY_GOODNESS_OF_FIT"
         == GraphAverageDegreeNormalityGofStatistic().code()
+    )
+
+
+@pytest.mark.parametrize(
+    ("data", "expected_result"),
+    [
+        ([0.713, 0.644, 2.625, 0.740, 0.501, 0.185, 0.982, 1.028, 1.152, 0.267], 6.0),
+        ([0.039, 3.036, 0.626, 1.107, 0.139, 1.629, 0.050, 0.118, 0.978, 2.699], 4.0),
+    ],
+)
+def test_clique_number_normality_criterion(data, expected_result):
+    statistic = GraphCliqueNumberNormalityGofStatistic().execute_statistic(data)
+    assert expected_result == pytest.approx(statistic, 0.00001)
+
+
+def test_clique_number_normality_criterion_code():
+    assert (
+        "CLIQUENUMBER_GRAPH_NORMALITY_GOODNESS_OF_FIT"
+        == GraphCliqueNumberNormalityGofStatistic.code()
+    )
+
+
+@pytest.mark.parametrize(
+    ("data", "expected_result"),
+    [
+        ([0.713, 0.644, 2.625, 0.740, 0.501, 0.185, 0.982, 1.028, 1.152, 0.267], 3.0),
+        ([0.039, 3.036, 0.626, 1.107, 0.139, 1.629, 0.050, 0.118, 0.978, 2.699], 6.0),
+    ],
+)
+def test_independence_number_normality_criterion(data, expected_result):
+    statistic = GraphIndependenceNumberNormalityGofStatistic().execute_statistic(data)
+    assert expected_result == pytest.approx(statistic, 0.00001)
+
+
+def test_independence_number_normality_criterion_code():
+    assert (
+        "INDEPENDENCENUMBER_GRAPH_NORMALITY_GOODNESS_OF_FIT"
+        == GraphIndependenceNumberNormalityGofStatistic.code()
     )
 
 
@@ -853,7 +898,7 @@ def test_number_of_components_criterion(data, result):
 
 def test_number_of_components_criterion_code():
     assert (
-        "ConnectedComponents_NORMALITY_GOODNESS_OF_FIT"
+        "CONNECTEDCOMPONENTS_GRAPH_NORMALITY_GOODNESS_OF_FIT"
         == GraphConnectedComponentsNormalityGofStatistic().code()
     )
 
